refactor: Move cluster apis into _cluster
Intermediate commit to let git keep track of the correct file renames
dues to a "cluster" directory already existing in apis/, but that
contains different files.
