+++ apis/_cluster/applications/v1alpha1/register.go
