+++ apis/_cluster/applications/v1alpha1/zz_generated.managedlist.go
