+++ apis/_cluster/applicationsets/v1alpha1/register.go
