+++ apis/_cluster/projects/v1alpha1/zz_generated.resolvers.go
@@ -20,8 +20,9 @@ package v1alpha1
 
 import (
 	"context"
-	v1alpha11 "github.com/crossplane-contrib/provider-argocd/apis/cluster/v1alpha1"
-	v1alpha1 "github.com/crossplane-contrib/provider-argocd/apis/repositories/v1alpha1"
+
+	v1alpha11 "github.com/crossplane-contrib/provider-argocd/apis/cluster/cluster/v1alpha1"
+	v1alpha1 "github.com/crossplane-contrib/provider-argocd/apis/cluster/repositories/v1alpha1"
 	reference "github.com/crossplane/crossplane-runtime/v2/pkg/reference"
 	errors "github.com/pkg/errors"
 	client "sigs.k8s.io/controller-runtime/pkg/client"
