+++ apis/_cluster/repositories/v1alpha1/zz_generated.managedlist.go
